requirements.txt
alembic==1.5.8
cffi == 1.14.5
click == 8.1.3
cryptography == 3.4.6
Flask == 2.1.2
Flask-CKEditor 
Flask-Login  == 0.5.0
Flask-Migrate == 2.5.3
Flask-SQLAlchemy  == 2.4.3
gunicorn == 20.1.0
Python-dotenv == 0.13.0
Psycopg2-binary == 2.8.5
Uwsgi == 2.0.18
Requests == 2.23.0
itsdangerous == 2.1.2
Jinja2 == 3.1.2
Mako == 1.1.4
MarkupSafe ==  2.1.1
mysql-connector == 2.2.9
mysql-connector-python == 8.0.23
protobuf == 3.15.5
pycparser == 2.20
PyMySQL == 1.0.2
python-dateutil == 2.8.1
python-editor == 1.0.4
six == 1.15.0
SQLAlchemy == 1.3.23
Werkzeug == 2.1.2
WTForms == 2.3.3
Turbo-Flask  == 0.8.0
pandas == 1.4.3

